+++ src/main/java/com/hetero/repository/TokenDao.java
@@ -16,7 +16,7 @@ public interface TokenDao extends JpaRepository<Token, Integer> {
     select t from Token t inner join User u on t.user.id = u.id
     where t.user.id = :userId and t.loggedOut = false
    """)
-    List<Token> findAllAccessTokensByUser(Integer userId);
+    List<Token> findAllAccessTokensByUser(Long userId);
 
     Optional<Token> findByAccessToken(String token);
