+++ src/main/java/com/hetero/service/UserService.java
@@ -9,20 +9,20 @@ import java.util.List;
 public interface UserService {
 
     User addUser(User user);
-    User updateUser(Integer id,User user);
-    String deleteUser(Integer id);
-    User getUser(Integer id);
+    User updateUser(Long id,User user);
+    String deleteUser(Long id);
+    User getUser(Long id);
 
     void updateUserTransactions(User user);
 
-    List<Transaction> getUserTransactions(Integer userId);
+    List<Transaction> getUserTransactions(Long userId);
     List<User> getAllUsers();
     List<User> getAllUnBlockedUsers();
     List<User> getAllBlockedUsers();
-    User updateUserBlockStatus(Integer id, User newUser);
+    User updateUserBlockStatus(Long id, User newUser);
 
-    void blockUser(Integer userId);
-    void unblockUser(Integer userId);
+    void blockUser(Long userId);
+    void unblockUser(Long userId);
 
     List<User> getUsersByPlatform(Platform platform);
     List<User> getUnblockedUsersByPlatform(Platform platform);
